feat: Add displayName to GSI and getRoutesByGym response
