+++ lambda_backend/database_setup/serverless.yml
@@ -46,6 +46,7 @@ resources:
                 - publicGrade
                 - voteCount
                 - commentCount
+                - displayName
               ProjectionType: INCLUDE
             ProvisionedThroughput:
               ReadCapacityUnits: ${self:custom._routeDataTableGSI.ReadCapacityUnits.${self:custom.stage}}
