+++ lambda_backend/route_microservice/src/createRoute.ts
@@ -1,18 +1,15 @@
 import { Handler } from 'aws-lambda';
-import DynamoDB, {
-  AttributeValue,
-  QueryInput,
-  PutItemInput,
-  AttributeMap,
-} from 'aws-sdk/clients/dynamodb';
+import DynamoDB, { PutItemInput, AttributeMap } from 'aws-sdk/clients/dynamodb';
 import S3, { PutObjectRequest } from 'aws-sdk/clients/s3';
 
 import {
   getMiddlewareAddedHandler,
+  getGymIsRegistered,
   CreateRouteEvent,
   createRouteSchema,
   JwtPayload,
   RouteItem,
+  getCognitoUserDetails,
 } from './common';
 import { MAX_PHOTO_SIZE } from './config';
 import { createHash } from 'crypto';
@@ -35,29 +32,6 @@ const createRoute: Handler = async (event: CreateRouteEvent) => {
     body: { countryCode, routeName, expiredTime, gymLocation, ownerGrade, routePhoto },
   } = event;
 
-  const queryInput: QueryInput = {
-    TableName: process.env['GYM_TABLE_NAME'],
-    ConsistentRead: false,
-    KeyConditionExpression: 'countryCode = :countryCode AND gymLocation = :gymLocation',
-    ExpressionAttributeValues: {
-      ':countryCode': countryCode as AttributeValue,
-      ':gymLocation': gymLocation as AttributeValue,
-    },
-  };
-  let Items;
-  try {
-    ({ Items } = await dynamoDb.query(queryInput).promise());
-  } catch (error) {
-    throw createError(500, 'Error querying table :' + error.stack);
-  }
-  if (Items.length === 0) {
-    return {
-      statusCode: 400,
-      body: JSON.stringify({ Message: 'Unregistered gym' }),
-    };
-  }
-
-  const { username } = (await jwt_decode(Authorization.split(' ')[1])) as JwtPayload;
   const { content, mimetype } = routePhoto;
   if (mimetype !== 'image/jpeg') {
     throw createError(400, 'Only JPEG is accepted');
@@ -65,6 +39,16 @@ const createRoute: Handler = async (event: CreateRouteEvent) => {
   if (content.byteLength > MAX_PHOTO_SIZE) {
     throw createError(400, 'Max allowed size is 5MB');
   }
+  const gymIsRegistered = await getGymIsRegistered(countryCode, gymLocation);
+  if (!gymIsRegistered) {
+    return {
+      statusCode: 400,
+      body: JSON.stringify({ Message: 'Unregistered gym' }),
+    };
+  }
+
+  const accessToken = Authorization.split(' ')[1];
+  const { username } = (await jwt_decode(accessToken)) as JwtPayload;
 
   const createdAt = new Date().toISOString();
   const usernameHash = createHash('sha256').update(username).digest('base64');
@@ -81,10 +65,13 @@ const createRoute: Handler = async (event: CreateRouteEvent) => {
     throw createError(500, 'S3 upload failed.' + error.stack);
   }
 
+  const displayName = (await getCognitoUserDetails(accessToken)).fullName;
+
   const routeItem: RouteItem = {
     username,
     createdAt,
     ttl: new Date(expiredTime).getTime(),
+    displayName,
     routeName,
     gymLocation,
     routeURL,
